chore: Split CRD and YAML specific code
