+++ crates/stackable-operator/src/lib.rs
@@ -22,7 +22,11 @@ pub mod utils;
 pub mod validation;
 
 // Internal re-exports
-pub use stackable_shared::yaml::CustomResourceExt;
+pub use stackable_shared::{crd::CustomResourceExt, yaml::YamlSchema};
+
+pub mod shared {
+    pub use stackable_shared::*;
+}
 
 // External re-exports
 pub use ::k8s_openapi;
