+++ crates/stackable-shared/src/lib.rs
@@ -1,4 +1,5 @@
 //! This crate contains various shared helpers and utilities used across other crates in this
 //! workspace.
 
+pub mod crd;
 pub mod yaml;
